+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.0.4] - 2021-02-21
 ### Fixed
 - Fix crash when `seo_schema_org` when both given entity and current entity are null
 - Fix crash when `$reflectionEntity` is `null` in `UrlHistoryWriter::loadClassMetadata`
@@ -33,7 +35,8 @@ Initial release for the v2.
 ## [1.2.6] - 2020-05-28
 Last release of the v1.
 
-[Unreleased]: https://github.com/umanit/block-collection-bundle/compare/2.0.3...HEAD
+[Unreleased]: https://github.com/umanit/block-collection-bundle/compare/2.0.4...HEAD
+[2.0.4]: https://github.com/umanit/block-collection-bundle/compare/2.0.3...2.0.4
 [2.0.3]: https://github.com/umanit/block-collection-bundle/compare/2.0.2...2.0.3
 [2.0.2]: https://github.com/umanit/block-collection-bundle/compare/2.0.1...2.0.2
 [2.0.1]: https://github.com/umanit/block-collection-bundle/compare/2.0.0...2.0.1